+++ R/plotGeneRegion.R
@@ -101,16 +101,28 @@ prepareGTF <- function(gtf, transcriptIdColumn = "transcript_id",
 #'   padding in the lower and upper range of the plot, respectively. For 
 #'   example, a value of 0.05 will expand the range by 
 #'   0.05 * (max coordinate - min coordinate) in the specified direction. 
+#' @param colorByStrand Logical scalar, determining whether gene features are 
+#'   colored by the annotated strand.
+#' @param featureColors Named character vector of length 4, with elements 
+#'   \code{plusmain}, \code{minusmain}, \code{plusother}, \code{minusother}, 
+#'   giving the colors to use for the features if \code{colorByStrand} is TRUE. 
+#' @param condColors Either NULL or a named character vector (with the same 
+#'   names as the unique values of \code{bigwigCond}), giving the colors to use
+#'   for the coverage tracks if \code{bigwigCond} is provided. 
+#' @param scaleDataTracks Logical scalar, indicating whether the data tracks 
+#'   should be scaled to have the same y-axis limits. 
 #' 
 #' @author Charlotte Soneson
 #' 
 #' @export
 #' 
-#' @importFrom BiocGenerics subset start end
+#' @importFrom BiocGenerics subset start end strand
 #' @importFrom GenomeInfoDb seqnames
-#' @importFrom IRanges overlapsAny IRanges
+#' @importFrom IRanges overlapsAny IRanges subsetByOverlaps
+#' @importFrom rtracklayer import
 #' @importFrom GenomicRanges GRanges
-#' @importFrom Gviz GeneRegionTrack DataTrack GenomeAxisTrack plotTracks
+#' @importFrom Gviz GeneRegionTrack DataTrack GenomeAxisTrack plotTracks 
+#'   feature
 #' @importFrom S4Vectors %in%
 #' @importFrom methods is
 #'   
@@ -144,7 +156,14 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
                            geneIdColumn = "gene_id",
                            geneSymbolColumn = "gene_name",
                            lowerPadding = 0.15, 
-                           upperPadding = 0.05) {
+                           upperPadding = 0.05,
+                           colorByStrand = FALSE, 
+                           featureColors = c(plusmain = "#0E14D0", 
+                                             minusmain = "#D0350E",
+                                             plusother = "#9E9BEB", 
+                                             minusother = "#DA907E"),
+                           condColors = NULL,
+                           scaleDataTracks = FALSE) {
     options(ucscChromosomeNames = FALSE)
     
     ## ---------------------------------------------------------------------- ##
@@ -197,6 +216,33 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
         !methods::is(upperPadding, "numeric") || 
         length(lowerPadding) != 1 || length(upperPadding) != 1) {
         stop("'lowerPadding' and 'upperPadding' must be numeric scalars")
+    }
+    if (!is.logical(colorByStrand) || 
+        length(colorByStrand) != 1) {
+        stop("'colorByStrand' must be a logical scalar")
+    }
+    if (!methods::is(featureColors, "character") || 
+        length(featureColors) != 4) {
+        stop("'featureColors' must be a character vector of length 4")
+    }
+    if (length(intersect(names(featureColors), 
+                         c("plusmain", "minusmain", "plusother", "minusother"))) != 4) {
+        stop("'featureColors must have elements 'plusmain', 'minusmain', ",
+             "'plusother', 'minusother'")
+    }
+    if (!is.null(condColors)) {
+        if (!is(condColors, "character")) {
+            stop("'condColors' must be a character vector or NULL")
+        }
+        if (any(bigwigCond != "") && 
+            length(intersect(unique(bigwigCond), names(condColors))) != 
+            length(condColors)) {
+            stop("'condColors' must be a named vector with names equal to the ",
+                 "values of bigwigCond")
+        }
+    }
+    if (length(scaleDataTracks) != 1 || !is.logical(scaleDataTracks)) {
+        stop("'scaleDataTracks' must be a logical scalar")
     }
     ## Must have at least one of bigwigFiles, gtf and granges
     if (all(bigwigFiles == "") && is.null(granges) && gtf == "") {
@@ -265,8 +311,24 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
         
         grtr <- Gviz::GeneRegionTrack(gr, showId = TRUE, col = NULL, fill = "gray80",
                                       name = geneTrackTitle, col.title = "black")
-        grtr2 <- Gviz::GeneRegionTrack(gro, showId = TRUE, col = "black", fill = "white",
-                                       name = "", col.title = "black")
+        
+        if (colorByStrand) {
+            grtr2 <- Gviz::GeneRegionTrack(gro, showId = TRUE, col = NULL, 
+                                           fill = "white",
+                                           name = "", col.title = "black")
+            if (length(grtr) > 0) {
+                Gviz::feature(grtr) <- ifelse(BiocGenerics::strand(grtr) == "+", 
+                                              "plusmain", "minusmain")
+            }
+            if (length(grtr2) > 0) {
+                Gviz::feature(grtr2) <- ifelse(BiocGenerics::strand(grtr2) == "+", 
+                                               "plusother", "minusother")
+            }
+        } else {
+            grtr2 <- Gviz::GeneRegionTrack(gro, showId = TRUE, col = "black", 
+                                           fill = "white",
+                                           name = "", col.title = "black")
+        }
     } else {
         gr <- gro <- grtr <- grtr2 <- NULL
     }    
@@ -296,23 +358,48 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
     ## character strings)
     if (!any(bigwigFiles == "")) {
         
+        ## Determine the y limits
+        ## Thanks to http://adomingues.github.io/2016/11/13/max-coverage-in-bigwigs/
+        ## for the idea
+        if (!scaleDataTracks) {
+            ylim <- NULL
+        } else {
+            g <- GenomicRanges::GRanges(seqnames = chr,
+                                        ranges = IRanges(start = minCoord,
+                                                         end = maxCoord),
+                                        strand = "*")
+            ymax <- max(vapply(bigwigFiles, function(f) {
+                ff <- rtracklayer::import(f, format = "bw", which = g)
+                if (length(ff) > 0) {
+                    return(max(ff$score))
+                } else {
+                    return(0)
+                }
+            }, 0))
+            ylim <- c(0, ymax)
+        }
+        
         ## ---------------------------------------------------------------------- ##
         ## Define colors if bigwigCond is provided
         ## ---------------------------------------------------------------------- ##
-        ## Define colors for coverage tracks
-        color_list <- rep(c("#DC050C", "#7BAFDE", "#B17BA6", "#F1932D", "#F7EE55",
-                            "#90C987", "#777777", "#E8601C", "#1965B0", "#882E72",
-                            "#F6C141", "#4EB265", "#CAEDAB"), 
-                          ceiling(length(unique(bigwigCond))/13))
-        
-        if (length(bigwigCond) > 1 || bigwigCond != "") {
-            usecol <- color_list[match(bigwigCond, 
-                                       unique(bigwigCond))]
+        if (!is.null(condColors)) {
+            usecol <- condColors
         } else {
-            usecol <- rep("gray", length(bigwigFiles))
-        }
-        names(usecol) <- names(bigwigCond)
+            ## Define colors for coverage tracks
+            color_list <- rep(c("#DC050C", "#7BAFDE", "#B17BA6", "#F1932D", "#F7EE55",
+                                "#90C987", "#777777", "#E8601C", "#1965B0", "#882E72",
+                                "#F6C141", "#4EB265", "#CAEDAB"), 
+                              ceiling(length(unique(bigwigCond))/13))
             
+            if (length(bigwigCond) > 1 || bigwigCond != "") {
+                usecol <- color_list[match(bigwigCond, 
+                                           unique(bigwigCond))]
+            } else {
+                usecol <- rep("gray", length(bigwigFiles))
+            }
+            names(usecol) <- names(bigwigCond)
+        }
+        
         ## ------------------------------------------------------------------ ##
         ## Prepare final plot
         ## ------------------------------------------------------------------ ##
@@ -326,7 +413,8 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
                                    fill = usecol[i],
                                    col = usecol[i],
                                    col.histogram = usecol[i],
-                                   fill.histogram = usecol[i]))
+                                   fill.histogram = usecol[i],
+                                   ylim = ylim))
         })
     } else {
         tracks <- NULL
@@ -341,5 +429,9 @@ plotGeneRegion <- function(gtf = "", granges = NULL, chr = "",
     Gviz::plotTracks(tracks, chromosome = chr, from = minCoord, 
                      to = maxCoord, main = plotTitle, 
                      transcriptAnnotation = "transcript",
-                     min.width = 0, min.distance = 0, collapse = FALSE)
+                     min.width = 0, min.distance = 0, collapse = FALSE,
+                     plusmain = featureColors["plusmain"], 
+                     minusmain = featureColors["minusmain"],
+                     plusother = featureColors["plusother"], 
+                     minusother = featureColors["minusother"])
 }