+++ NEWS.md
@@ -1,3 +1,11 @@
+# swissknife 0.15
+
+* Allow specification of colors for coverage tracks in plotGeneRegion
+
+# swissknife 0.14
+
+* Allow coloring features by strand in plotGeneRegion
+
 # swissknife 0.13
 
 * Add parsePkgVersions